modify style issue
