+++ src/components/Dropdown/index.js
@@ -3,7 +3,7 @@ import {NavLink} from "react-router-dom";
 import "./styles.css";
 import  UserContext from "../../context/Context";
 
-export default function Dropdown() {  
+export default function Dropdown(props) {  
   const [isActive, setIsActive] = useState(false);
 
   const onClickHandle = (e) => {
@@ -15,6 +15,11 @@ export default function Dropdown() {
     setIsActive(false);
   }
 
+  const logoutHalde = () => {
+    setIsActive(false);
+    props.onClick();
+  }
+
   const {user, setUser}  = useContext(UserContext);
 
   return (
@@ -22,18 +27,21 @@ export default function Dropdown() {
       <div className="menu-container">
         <button onClick={(e) => onClickHandle(e)} className="menu-trigger">
           <span>{user.surname}</span>
-        </button>
-        <nav
-          
+          <i className = {`fas fa-caret-down ${isActive ? "active" : "inactive"}`}></i>
+        </button>        
+        <nav          
           className={`menu ${isActive ? "active" : "inactive"}`}
         >
           <ul>
             <li>              
-              <NavLink to={'/user-profile'} onClick = {showDropdown}>Manage Profile</NavLink>
+              <NavLink to={'/user-profile'} onClick = {showDropdown}>MANAGE PROFILE</NavLink>
             </li>
             <li>              
-              <NavLink to={'/manage-ads'} onClick = {showDropdown}>Manage Ads</NavLink>
-            </li>            
+              <NavLink to={'/manage-ads'} onClick = {showDropdown}>MANAGE ADS</NavLink>
+            </li>   
+            <li>              
+              <NavLink to={'/'} onClick = {logoutHalde}>LOGOUT</NavLink>
+            </li>          
           </ul>
         </nav>
       </div>
