+++ src/components/Dropdown/styles.css
@@ -21,9 +21,9 @@ body {
     display: flex;
     justify-content: space-between;
     align-items: center;
-    padding: 4px 6px;
+    padding: 5px 10px 5px 5px;
     box-shadow: 0 1px 3px rgba(0, 0, 0, 0.3);
-    border: none;
+    border: 1px solid !important;
     vertical-align: middle;
     transition: box-shadow 0.4s ease;
     margin-left: auto; /* Strictly for positioning */
@@ -46,7 +46,7 @@ body {
 
 .menu {
     z-index: 11;
-    background: #ffffff;
+    background: #161e21;
     border-radius: 8px;
     position: absolute;
     top: 30px;
@@ -64,6 +64,17 @@ body {
     visibility: visible;
     transform: translateY(0);
     z-index: 11;
+    background-color: #161e21;
+}
+
+.fa-caret-down.active {
+    transform: rotate(180deg);
+    transition: opacity 0.4s ease, transform 0.4s ease, visibility 0.4s;
+}
+
+.fa-caret-down.inactive {
+    transform: rotate(0deg);
+    transition: opacity 0.4s ease, transform 0.4s ease, visibility 0.4s;
 }
 
 .menu ul {
@@ -78,7 +89,11 @@ body {
 
 .menu li a {
     text-decoration: none;
-    color: #333333;
+    color: white;
     padding: 15px 20px;
     display: block;
+}
+
+.fa-caret-down {
+    font-size: 20px;
 }
