+++ src/firebase/User.js
@@ -1,10 +1,16 @@
 import {doc, setDoc, getDoc, db} from './index';
 
-export function AddUser(docID, data) { 
-    return setDoc(doc(db, "users", docID), data);
+const collection_user = "users";
+
+export const saveUser = async (docID, surname) => {
+    const data = {
+        roles: ["CUSTOMER"],
+        phone: "1234567890",
+        surname: surname
+    }
+    await setDoc(doc(db, collection_user, docID), data);
 }
 
-export function GetUser(id) {
-    return getDoc(doc(db, "users", id));
+export const getUser = (id) => {
+    return getDoc(doc(db, collection_user, id));
 }
-
